+++ my.h
@@ -9,7 +9,7 @@
 #ifndef MY_H_
 #define MY_H_
 void move(node_t **dest, node_t **src);
-void append(node_t **head, int val);
+void append(node_t **head, long int val);
 void printlist(node_t *head);
 int move_end(node_t **head_ref);
 int radix_binary(node_t *l_a, node_t *l_b, int neg);
@@ -20,5 +20,5 @@ int is_sorted(node_t *l_a);
 int my_putstr(char const *str);
 int get_size(node_t *l_a);
 void my_putchar (char c);
-int my_get_nbr(char *str);
+long int my_get_nbr(char *str);
 #endif /* !MY_H_ */
